pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbt-cost"
version = "0.1.0"
description = "Estimate BigQuery costs for dbt models before execution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Ishan Yash" },
]
keywords = ["dbt", "bigquery", "cost", "estimation", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "dbt-lens>=0.3.0",
    "click>=8.0",
    "rich>=13.0",
    "google-cloud-bigquery>=3.0",
]

[project.scripts]
dbt-cost = "dbt_cost.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dbt_cost"]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
